refactor(swagger-ui): update server module layout (#118)
