+++ swagger-ui-3/Dockerfile
@@ -1,5 +1,7 @@
 FROM registry.opensource.zalan.do/stups/node:7.10-cd52
 
+ARG conf=conf/production.yaml
+
 RUN apt-get update && apt-get install -y apt-transport-https \
     && curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | apt-key add - \
     && echo "deb https://dl.yarnpkg.com/debian/ stable main" | tee /etc/apt/sources.list.d/yarn.list \
@@ -8,12 +10,15 @@ RUN apt-get update && apt-get install -y apt-transport-https \
 
 ADD package.json /opt/api-discovery-ui/
 ADD yarn.lock /opt/api-discovery-ui/
+
+WORKDIR /opt/api-discovery-ui
+RUN yarn install --production
+
 ADD dist/client /opt/api-discovery-ui/client
 ADD dist/server /opt/api-discovery-ui/server
 
-WORKDIR /opt/api-discovery-ui
-
-RUN yarn install --production
+ENV API_DISCOVERY_CONF /opt/api-discovery-ui/conf.yaml
+ADD $conf $API_DISCOVERY_CONF
 
 ENV NODE_ENV production
