+++ swagger-ui-3/conf/development.yaml
@@ -1,0 +1,3 @@
+# Development configuration file.
+logLevel: debug
+enableWebpackDev: true
