+++ swagger-ui-3/conf/production.yaml
@@ -1,0 +1,2 @@
+# Production configuration file (example)
+logLevel: info
