+++ swagger-ui-3/package.json
@@ -10,8 +10,8 @@
     "build-server": "tsc -p .",
     "build": "npm run build-server && npm run build-client",
     "dist": "npm run clean && npm run build-server && NODE_ENV=production npm run build-client -- -p",
-    "start-dist": "NODE_ENV=production node dist/server/index.js",
-    "start": "API_DISCOVERY_ENABLE_WEBPACK_DEV=true nodemon server/index.ts",
+    "start-dist": "NODE_ENV=production API_DISCOVERY_CONF=conf/production.yaml node dist/server/index.js",
+    "start": "API_DISCOVERY_CONF=conf/development.yaml nodemon server/index.ts",
     "lint": "eslint client && tslint --type-check --project tsconfig.json",
     "test": "npm run lint && API_DISCOVERY_LOG_LEVEL=warn jest --coverage",
     "prettier": "prettier --print-width=100 --single-quote --write \"{client,server}/**/*.+(js|ts)\""
@@ -24,7 +24,7 @@
     "@types/koa-router": "7.0.22",
     "@types/koa-static": "3.0.0",
     "@types/mz": "0.0.31",
-    "@types/node": "7.0.28",
+    "@types/node": "7.0.29",
     "@types/request-promise-native": "1.0.5",
     "@types/supertest": "2.0.0",
     "@types/winston": "2.3.3",
@@ -42,13 +42,13 @@
     "eslint-plugin-prettier": "2.1.1",
     "eslint-plugin-promise": "3.5.0",
     "eslint-plugin-react": "7.0.1",
-    "extract-text-webpack-plugin": "2.1.0",
+    "extract-text-webpack-plugin": "2.1.2",
     "file-loader": "0.11.2",
     "html-webpack-plugin": "2.28.0",
     "isomorphic-fetch": "2.2.1",
     "jest": "20.0.4",
     "nodemon": "1.11.0",
-    "prettier": "1.4.2",
+    "prettier": "1.4.4",
     "prop-types": "15.5.10",
     "react": "15.5.4",
     "react-dom": "15.5.4",
@@ -72,7 +72,7 @@
     "koa-static": "3.0.0",
     "mz": "2.6.0",
     "request-promise-native": "1.0.4",
-    "typeconf": "0.1.1",
+    "typeconf": "0.1.2",
     "winston": "2.3.1"
   },
   "jest": {
