+++ swagger-ui-3/server/domain/health.ts
@@ -1,0 +1,5 @@
+import { Context } from 'koa';
+
+export default async function(ctx: Context) {
+  ctx.body = 'OK';
+}
