+++ swagger-ui-3/server/framework/webpack-dev.js
@@ -1,5 +1,5 @@
 const webpack = require('webpack');
 const webpackDev = require('webpack-koa2-middleware');
-const webpackConfig = require('../webpack.config');
+const webpackConfig = require('../../webpack.config');
 
 module.exports = options => webpackDev(webpack(webpackConfig), options);
