+++ swagger-ui-3/server/resource.ts
@@ -1,0 +1,6 @@
+import Router = require('koa-router');
+import health from './domain/health';
+
+const createRoutes = (router: Router) => router.get('/health', health);
+
+export default () => createRoutes(new Router());
