--- swagger-ui-3/server/routes/health.ts
@@ -1,6 +1,0 @@
-import Router = require('koa-router');
-
-export default (router: Router) =>
-  router.get('/health', ctx => {
-    ctx.body = 'OK';
-  });
